@@ -15,8 +15,8 @@
 
 [tool.poetry]
 name = "neo4j-graphrag"
-version = "0.6.1"
-description = "Python package to allow easy integration to Neo4j's GenAI features"
+version = "1.1.0"
+description = "Python package to allow easy integration to Neo4j's GraphRAG features"
 authors = ["Neo4j, Inc <team-gen-ai@neo4j.com>"]
 license = "Apache License, Version 2.0"
 readme = "README.md"
@@ -28,19 +28,25 @@ include = "neo4j_graphrag"
 from = "src"
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.9.0"
 neo4j = "^5.17.0"
-types-requests = "^2.31.0.20240218"
 pydantic = "^2.6.3"
+urllib3 = "<2"
 weaviate-client = {version = "^4.6.1", optional = true}
 pinecone-client = {version = "^4.1.0", optional = true}
 types-mock = "^5.1.0.20240425"
 eval-type-backport = "^0.2.0"
 pypdf = "^4.3.1"
+fsspec = "^2024.9.0"
 pygraphviz = [
     {version = "^1.13.0", python = ">=3.10,<4.0.0", optional = true},
     {version = "^1.0.0", python = "<3.10", optional = true}
 ]
+google-cloud-aiplatform = {version = "^1.66.0", optional = true}
+cohere = {version = "^5.9.0", optional = true}
+anthropic = { version = "^0.34.2", optional = true}
+mistralai = {version = "^1.0.3", optional = true}
+qdrant-client = {version = "^1.11.3", optional = true}
 
 [tool.poetry.group.dev.dependencies]
 pylint = "^3.1.0"
@@ -50,14 +56,12 @@ pytest-mock = "^3.12.0"
 pre-commit = { version = "^3.6.2", python = "^3.9" }
 coverage = "^7.4.3"
 ruff = "^0.3.0"
-langchain-openai = "^0.1.1"
+langchain-text-splitters = "^0.3.0"
 weaviate-client = "^4.6.1"
 sentence-transformers = "^3.0.0"
 pinecone-client = "^4.1.0"
-langchain-community = "^0.2.0"
 requests = "^2.32.0"
 sphinx = { version = "^7.2.6", python = "^3.9" }
-langchain-huggingface = "^0.0.3"
 tox = "^4.15.1"
 numpy = [
     {version = "^1.24.0", python = "<3.12"},
@@ -73,9 +77,15 @@ pygraphviz = [
     {version = "^1.13.0", python = ">=3.10,<4.0.0"},
     {version = "^1.0.0", python = "<3.10"}
 ]
+google-cloud-aiplatform = {version = "^1.66.0"}
+cohere = {version = "^5.9.0"}
+anthropic = { version = "^0.34.2"}
+mistralai = {version = "^1.0.3"}
+qdrant-client = {version = "^1.11.3"}
+langchain-openai = "^0.2.2"  # needed in the examples
 
 [tool.poetry.extras]
-external_clients = ["weaviate-client", "pinecone-client"]
+external_clients = ["weaviate-client", "pinecone-client", "google-cloud-aiplatform", "cohere", "anthropic", "mistralai", "qdrant-client"]
 kg_creation_tools = ["pygraphviz"]
 
 [build-system]
