@@ -85,13 +85,18 @@ class GraphRAG:
         query_text: str = "",
         examples: str = "",
         retriever_config: Optional[dict[str, Any]] = None,
-        return_context: bool = False,
-        query: Optional[str] = None,
+        return_context: bool | None = None,
     ) -> RagResultModel:
-        """This method performs a full RAG search:
-        1. Retrieval: context retrieval
-        2. Augmentation: prompt formatting
-        3. Generation: answer generation with LLM
+        """
+        .. warning::
+            The default value of 'return_context' will change from 'False' to 'True' in a future version.
+
+
+        This method performs a full RAG search:
+            1. Retrieval: context retrieval
+            2. Augmentation: prompt formatting
+            3. Generation: answer generation with LLM
+
 
         Args:
             query_text (str): The user question
@@ -99,28 +104,18 @@ class GraphRAG:
             retriever_config (Optional[dict]): Parameters passed to the retriever
                 search method; e.g.: top_k
             return_context (bool): Whether to append the retriever result to the final result (default: False)
-            query (Optional[str]): The user question. Will be deprecated in favor of query_text.
 
         Returns:
             RagResultModel: The LLM-generated answer
 
         """
+        if return_context is None:
+            warnings.warn(
+                "The default value of 'return_context' will change from 'False' to 'True' in a future version.",
+                DeprecationWarning,
+            )
+            return_context = False
         try:
-            if query is not None:
-                if query_text:
-                    warnings.warn(
-                        "Both 'query' and 'query_text' are provided, 'query_text' will be used.",
-                        DeprecationWarning,
-                        stacklevel=2,
-                    )
-                elif isinstance(query, str):
-                    warnings.warn(
-                        "'query' is deprecated and will be removed in a future version, please use 'query_text' instead.",
-                        DeprecationWarning,
-                        stacklevel=2,
-                    )
-                    query_text = query
-
             validated_data = RagSearchModel(
                 query_text=query_text,
                 examples=examples,
