@@ -24,7 +24,11 @@ from datetime import datetime
 from timeit import default_timer
 from typing import Any, AsyncGenerator, Optional
 
-import pygraphviz as pgv
+try:
+    import pygraphviz as pgv
+except ImportError:
+    pgv = None
+
 from pydantic import BaseModel, Field
 
 from neo4j_graphrag.experimental.pipeline.component import Component, DataModel
@@ -191,8 +195,9 @@ class Orchestrator:
         for d in dependencies:
             d_status = await self.get_status_for_component(d.start)
             if d_status != RunStatus.DONE:
-                logger.warning(
-                    f"Missing dependency {d.start} for {task.name} (status: {d_status})"
+                logger.debug(
+                    f"Missing dependency {d.start} for {task.name} (status: {d_status}). "
+                    "Will try again when dependency is complete."
                 )
                 raise PipelineMissingDependencyError()
 
@@ -386,6 +391,12 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
         G.draw(path)
 
     def get_pygraphviz_graph(self, hide_unused_outputs: bool = True) -> pgv.AGraph:
+        if pgv is None:
+            raise ImportError(
+                "Could not import pygraphviz. "
+                "Follow installation instruction in pygraphviz documentation "
+                "to get it up and running on your system."
+            )
         self.validate_parameter_mapping()
         G = pgv.AGraph(strict=False, directed=True)
         # create a node for each component
@@ -430,7 +441,7 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
         task = TaskPipelineNode(name, component)
         self.add_node(task)
         # invalidate the pipeline if it was already validated
-        self.is_validated = False
+        self.invalidate()
 
     def set_component(self, name: str, component: Component) -> None:
         """Replace a component with another. If 'name' is not yet in the pipeline,
@@ -439,7 +450,7 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
         task = TaskPipelineNode(name, component)
         self.set_node(task)
         # invalidate the pipeline if it was already validated
-        self.is_validated = False
+        self.invalidate()
 
     def connect(
         self,
@@ -475,7 +486,12 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
         if self.is_cyclic():
             raise PipelineDefinitionError("Cyclic graph are not allowed")
         # invalidate the pipeline if it was already validated
+        self.invalidate()
+
+    def invalidate(self) -> None:
         self.is_validated = False
+        self.param_mapping = defaultdict(dict)
+        self.missing_inputs = defaultdict()
 
     def validate_parameter_mapping(self) -> None:
         """Go through the graph and make sure parameter mapping is valid
@@ -520,6 +536,8 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
 
         Considering the naming {param => target (component, [output_parameter]) },
         the mapping is valid if:
+         - 'param' is a valid input for task
+         - 'param' has not already been mapped
          - The target component exists in the pipeline and, if specified, the
             target output parameter is a valid field in the target component's
             result model.
@@ -543,6 +561,14 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
             # check that the previous component is actually returning
             # the mapped parameter
             for param, path in edge_inputs.items():
+                if param in self.param_mapping[task.name]:
+                    raise PipelineDefinitionError(
+                        f"Parameter '{param}' already mapped to {self.param_mapping[task.name][param]}"
+                    )
+                if param not in task.component.component_inputs:
+                    raise PipelineDefinitionError(
+                        f"Parameter '{param}' is not a valid input for component '{task.name}' of type '{task.component.__class__.__name__}'"
+                    )
                 try:
                     source_component_name, param_name = path.split(".")
                 except ValueError:
@@ -580,6 +606,7 @@ class Pipeline(PipelineGraph[TaskPipelineNode, PipelineEdge]):
     async def run(self, data: dict[str, Any]) -> PipelineResult:
         logger.debug("Starting pipeline")
         start_time = default_timer()
+        self.invalidate()
         self.validate_input_data(data)
         orchestrator = Orchestrator(self)
         await orchestrator.run(data)
