@@ -17,9 +17,11 @@ import re
 from typing import Any, Generator
 
 import pytest
-from langchain_huggingface.embeddings import HuggingFaceEmbeddings
 from neo4j import Driver
-from neo4j_graphrag.embedder import Embedder
+from neo4j_graphrag.embeddings.base import Embedder
+from neo4j_graphrag.embeddings.sentence_transformers import (
+    SentenceTransformerEmbeddings,
+)
 from neo4j_graphrag.retrievers import WeaviateNeo4jRetriever
 from neo4j_graphrag.types import RetrieverResult, RetrieverResultItem
 from weaviate.client import WeaviateClient
@@ -30,8 +32,10 @@ from .populate_dbs import populate_dbs
 
 
 @pytest.fixture(scope="module")
-def sentence_transformer_embedder() -> Generator[HuggingFaceEmbeddings, Any, Any]:
-    embedder = HuggingFaceEmbeddings(model_name="all-MiniLM-L6-v2")
+def sentence_transformer_embedder() -> (
+    Generator[SentenceTransformerEmbeddings, Any, Any]
+):
+    embedder = SentenceTransformerEmbeddings(model="all-MiniLM-L6-v2")
     yield embedder
 
 
