@@ -14,30 +14,76 @@
 #  limitations under the License.
 from __future__ import annotations
 
-from unittest.mock import MagicMock
+from unittest import mock
+from unittest.mock import MagicMock, Mock
 
 import pytest
-from neo4j_graphrag.experimental.components.kg_writer import Neo4jWriter
+from neo4j_graphrag.experimental.components.kg_writer import Neo4jWriter, batched
 from neo4j_graphrag.experimental.components.types import (
     Neo4jGraph,
     Neo4jNode,
     Neo4jRelationship,
 )
-from neo4j_graphrag.neo4j_queries import UPSERT_NODE_QUERY, UPSERT_RELATIONSHIP_QUERY
+from neo4j_graphrag.neo4j_queries import (
+    UPSERT_NODE_QUERY,
+    UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE,
+    UPSERT_RELATIONSHIP_QUERY,
+    UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE,
+)
 
 
-def test_upsert_node(driver: MagicMock) -> None:
+def test_batched() -> None:
+    assert list(batched([1, 2, 3, 4], batch_size=2)) == [
+        [1, 2],
+        [3, 4],
+    ]
+    assert list(batched([1, 2, 3], batch_size=2)) == [
+        [1, 2],
+        [3],
+    ]
+    assert list(batched([1, 2, 3], batch_size=4)) == [
+        [1, 2, 3],
+    ]
+
+
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._get_version",
+    return_value=(5, 22, 0),
+)
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+def test_upsert_nodes(_: Mock, driver: MagicMock) -> None:
     neo4j_writer = Neo4jWriter(driver=driver)
     node = Neo4jNode(id="1", label="Label", properties={"key": "value"})
-    neo4j_writer._upsert_node(node=node)
-    parameters = {"id": "1", "key": "value"}
+    neo4j_writer._upsert_nodes(nodes=[node])
     driver.execute_query.assert_called_once_with(
-        UPSERT_NODE_QUERY.format(label="Label", properties="{id: $id, key: $key}"),
-        parameters_=parameters,
+        UPSERT_NODE_QUERY,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {"key": "value"},
+                    "embedding_properties": None,
+                }
+            ]
+        },
     )
 
 
-def test_upsert_node_with_embedding(
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._get_version",
+    return_value=(5, 22, 0),
+)
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+def test_upsert_nodes_with_embedding(
+    _: Mock,
     driver: MagicMock,
 ) -> None:
     neo4j_writer = Neo4jWriter(driver=driver)
@@ -47,29 +93,33 @@ def test_upsert_node_with_embedding(
         properties={"key": "value"},
         embedding_properties={"embeddingProp": [1.0, 2.0, 3.0]},
     )
-    driver.execute_query.return_value.records = [{"elementID(n)": 1}]
-    neo4j_writer._upsert_node(node=node)
-    parameters = {"id": "1", "key": "value"}
+    driver.execute_query.return_value.records = [{"elementId(n)": 1}]
+    neo4j_writer._upsert_nodes(nodes=[node])
     driver.execute_query.assert_any_call(
-        UPSERT_NODE_QUERY.format(label="Label", properties="{id: $id, key: $key}"),
-        parameters_=parameters,
+        UPSERT_NODE_QUERY,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {"key": "value"},
+                    "embedding_properties": {"embeddingProp": [1.0, 2.0, 3.0]},
+                }
+            ]
+        },
     )
-    query = (
-        "MATCH (n) "
-        "WHERE elementId(n) = $id "
-        "WITH n "
-        "CALL db.create.setNodeVectorProperty(n, $embedding_property, $vector) "
-        "RETURN n"
-    )
-    parameters_ = {
-        "id": 1,
-        "embedding_property": "embeddingProp",
-        "vector": [1.0, 2.0, 3.0],
-    }
-    driver.execute_query.assert_any_call(query, parameters_, database_=None)
 
 
-def test_upsert_relationship(driver: MagicMock) -> None:
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._get_version",
+    return_value=(5, 22, 0),
+)
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+def test_upsert_relationship(_: Mock, driver: MagicMock) -> None:
     neo4j_writer = Neo4jWriter(driver=driver)
     rel = Neo4jRelationship(
         start_node_id="1",
@@ -77,18 +127,33 @@ def test_upsert_relationship(driver: MagicMock) -> None:
         type="RELATIONSHIP",
         properties={"key": "value"},
     )
-    neo4j_writer._upsert_relationship(rel=rel)
-    parameters = {"start_node_id": "1", "end_node_id": "2", "key": "value"}
+    neo4j_writer._upsert_relationships(rels=[rel])
+    parameters = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {"key": "value"},
+                "embedding_properties": None,
+            }
+        ]
+    }
     driver.execute_query.assert_called_once_with(
-        UPSERT_RELATIONSHIP_QUERY.format(
-            type="RELATIONSHIP",
-            properties="{key: $key}",
-        ),
+        UPSERT_RELATIONSHIP_QUERY,
         parameters_=parameters,
     )
 
 
-def test_upsert_relationship_with_embedding(driver: MagicMock) -> None:
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._get_version",
+    return_value=(5, 22, 0),
+)
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+def test_upsert_relationship_with_embedding(_: Mock, driver: MagicMock) -> None:
     neo4j_writer = Neo4jWriter(driver=driver)
     rel = Neo4jRelationship(
         start_node_id="1",
@@ -97,45 +162,305 @@ def test_upsert_relationship_with_embedding(driver: MagicMock) -> None:
         properties={"key": "value"},
         embedding_properties={"embeddingProp": [1.0, 2.0, 3.0]},
     )
-    driver.execute_query.return_value.records = [{"elementID(r)": "rel_elem_id"}]
-    neo4j_writer._upsert_relationship(rel=rel)
-    parameters = {"start_node_id": "1", "end_node_id": "2", "key": "value"}
+    driver.execute_query.return_value.records = [{"elementId(r)": "rel_elem_id"}]
+    neo4j_writer._upsert_relationships(rels=[rel])
+    parameters = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {"key": "value"},
+                "embedding_properties": {"embeddingProp": [1.0, 2.0, 3.0]},
+            }
+        ]
+    }
     driver.execute_query.assert_any_call(
-        UPSERT_RELATIONSHIP_QUERY.format(
-            type="RELATIONSHIP",
-            properties="{key: $key}",
-        ),
+        UPSERT_RELATIONSHIP_QUERY,
         parameters_=parameters,
     )
-    query = (
-        "MATCH ()-[r]->() "
-        "WHERE elementId(r) = $id "
-        "WITH r "
-        "CALL db.create.setRelationshipVectorProperty(r, $embedding_property, $vector) "
-        "RETURN r"
-    )
-    parameters_ = {
-        "id": "rel_elem_id",
-        "embedding_property": "embeddingProp",
-        "vector": [1.0, 2.0, 3.0],
-    }
-    driver.execute_query.assert_any_call(query, parameters_, database_=None)
 
 
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._get_version",
+    return_value=(5, 22, 0),
+)
 @pytest.mark.asyncio
-async def test_run(driver: MagicMock) -> None:
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+async def test_run(_: Mock, driver: MagicMock) -> None:
     neo4j_writer = Neo4jWriter(driver=driver)
     node = Neo4jNode(id="1", label="Label")
     rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
     graph = Neo4jGraph(nodes=[node], relationships=[rel])
     await neo4j_writer.run(graph=graph)
-    parameters = {"id": "1"}
     driver.execute_query.assert_any_call(
-        UPSERT_NODE_QUERY.format(label="Label", properties="{id: $id}"),
-        parameters_=parameters,
+        UPSERT_NODE_QUERY,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {},
+                    "embedding_properties": None,
+                }
+            ]
+        },
     )
-    parameters_ = {"start_node_id": "1", "end_node_id": "2"}
+    parameters_ = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {},
+                "embedding_properties": None,
+            }
+        ]
+    }
     driver.execute_query.assert_any_call(
-        UPSERT_RELATIONSHIP_QUERY.format(type="RELATIONSHIP", properties="{}"),
+        UPSERT_RELATIONSHIP_QUERY,
+        parameters_=parameters_,
+    )
+
+
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._get_version",
+    return_value=(5, 22, 0),
+)
+@pytest.mark.asyncio
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._async_db_setup",
+    return_value=None,
+)
+async def test_run_async_driver(_: Mock, async_driver: MagicMock) -> None:
+    neo4j_writer = Neo4jWriter(driver=async_driver)
+    node = Neo4jNode(id="1", label="Label")
+    rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
+    graph = Neo4jGraph(nodes=[node], relationships=[rel])
+    await neo4j_writer.run(graph=graph)
+    async_driver.execute_query.assert_any_call(
+        UPSERT_NODE_QUERY,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {},
+                    "embedding_properties": None,
+                }
+            ]
+        },
+    )
+    parameters_ = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {},
+                "embedding_properties": None,
+            }
+        ]
+    }
+    async_driver.execute_query.assert_any_call(
+        UPSERT_RELATIONSHIP_QUERY,
+        parameters_=parameters_,
+    )
+
+
+@pytest.mark.asyncio
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+async def test_run_is_version_below_5_23(_: Mock) -> None:
+    driver = MagicMock()
+    driver.execute_query = Mock(return_value=([{"versions": ["5.22.0"]}], None, None))
+
+    neo4j_writer = Neo4jWriter(driver=driver)
+
+    node = Neo4jNode(id="1", label="Label")
+    rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
+    graph = Neo4jGraph(nodes=[node], relationships=[rel])
+    await neo4j_writer.run(graph=graph)
+
+    driver.execute_query.assert_any_call(
+        UPSERT_NODE_QUERY,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {},
+                    "embedding_properties": None,
+                }
+            ]
+        },
+    )
+    parameters_ = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {},
+                "embedding_properties": None,
+            }
+        ]
+    }
+    driver.execute_query.assert_any_call(
+        UPSERT_RELATIONSHIP_QUERY,
+        parameters_=parameters_,
+    )
+
+
+@pytest.mark.asyncio
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._db_setup",
+    return_value=None,
+)
+async def test_run_is_version_5_23_or_above(_: Mock) -> None:
+    driver = MagicMock()
+    driver.execute_query = Mock(return_value=([{"versions": ["5.23.0"]}], None, None))
+
+    neo4j_writer = Neo4jWriter(driver=driver)
+    neo4j_writer.is_version_5_23_or_above = True
+
+    node = Neo4jNode(id="1", label="Label")
+    rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
+    graph = Neo4jGraph(nodes=[node], relationships=[rel])
+    await neo4j_writer.run(graph=graph)
+
+    driver.execute_query.assert_any_call(
+        UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {},
+                    "embedding_properties": None,
+                }
+            ]
+        },
+    )
+    parameters_ = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {},
+                "embedding_properties": None,
+            }
+        ]
+    }
+    driver.execute_query.assert_any_call(
+        UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE,
+        parameters_=parameters_,
+    )
+
+
+@pytest.mark.asyncio
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._async_db_setup",
+    return_value=None,
+)
+async def test_run_async_driver_is_version_below_5_23(_: Mock) -> None:
+    async_driver = MagicMock()
+    async_driver.execute_query = Mock(
+        return_value=([{"versions": ["5.22.0"]}], None, None)
+    )
+
+    neo4j_writer = Neo4jWriter(driver=async_driver)
+
+    node = Neo4jNode(id="1", label="Label")
+    rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
+    graph = Neo4jGraph(nodes=[node], relationships=[rel])
+    await neo4j_writer.run(graph=graph)
+
+    async_driver.execute_query.assert_any_call(
+        UPSERT_NODE_QUERY,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {},
+                    "embedding_properties": None,
+                }
+            ]
+        },
+    )
+    parameters_ = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {},
+                "embedding_properties": None,
+            }
+        ]
+    }
+    async_driver.execute_query.assert_any_call(
+        UPSERT_RELATIONSHIP_QUERY,
+        parameters_=parameters_,
+    )
+
+
+@pytest.mark.asyncio
+@mock.patch(
+    "neo4j_graphrag.experimental.components.kg_writer.Neo4jWriter._async_db_setup",
+    return_value=None,
+)
+async def test_run_async_driver_is_version_5_23_or_above(_: Mock) -> None:
+    async_driver = MagicMock()
+    async_driver.execute_query = Mock(
+        return_value=([{"versions": ["5.23.0"]}], None, None)
+    )
+
+    neo4j_writer = Neo4jWriter(driver=async_driver)
+
+    node = Neo4jNode(id="1", label="Label")
+    rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
+    graph = Neo4jGraph(nodes=[node], relationships=[rel])
+    await neo4j_writer.run(graph=graph)
+
+    async_driver.execute_query.assert_any_call(
+        UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE,
+        parameters_={
+            "rows": [
+                {
+                    "label": "Label",
+                    "labels": ["Label", "__Entity__"],
+                    "id": "1",
+                    "properties": {},
+                    "embedding_properties": None,
+                }
+            ]
+        },
+    )
+    parameters_ = {
+        "rows": [
+            {
+                "type": "RELATIONSHIP",
+                "start_node_id": "1",
+                "end_node_id": "2",
+                "properties": {},
+                "embedding_properties": None,
+            }
+        ]
+    }
+    async_driver.execute_query.assert_any_call(
+        UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE,
         parameters_=parameters_,
     )
