@@ -87,6 +87,7 @@ def test_hybrid_retriever_with_result_format_function(
         embedder,
         result_formatter=result_formatter,
     )
+    retriever.neo4j_version_is_5_23_or_above = True
     retriever.driver.execute_query.return_value = [  # type: ignore
         [neo4j_record],
         None,
@@ -174,12 +175,16 @@ def test_hybrid_search_text_happy_path(
     retriever = HybridRetriever(
         driver, vector_index_name, fulltext_index_name, embedder
     )
+    retriever.neo4j_version_is_5_23_or_above = True
     retriever.driver.execute_query.return_value = [  # type: ignore
         [neo4j_record],
         None,
         None,
     ]
-    search_query, _ = get_search_query(SearchType.HYBRID)
+    search_query, _ = get_search_query(
+        SearchType.HYBRID,
+        neo4j_version_is_5_23_or_above=retriever.neo4j_version_is_5_23_or_above,
+    )
 
     records = retriever.search(query_text=query_text, top_k=top_k)
 
@@ -226,12 +231,16 @@ def test_hybrid_search_favors_query_vector_over_embedding_vector(
         embedder,
         neo4j_database=database,
     )
+    retriever.neo4j_version_is_5_23_or_above = True
     retriever.driver.execute_query.return_value = [  # type: ignore
         [neo4j_record],
         None,
         None,
     ]
-    search_query, _ = get_search_query(SearchType.HYBRID)
+    search_query, _ = get_search_query(
+        SearchType.HYBRID,
+        neo4j_version_is_5_23_or_above=retriever.neo4j_version_is_5_23_or_above,
+    )
 
     retriever.search(query_text=query_text, query_vector=query_vector, top_k=top_k)
 
@@ -300,12 +309,17 @@ def test_hybrid_retriever_return_properties(
         embedder,
         return_properties,
     )
+    retriever.neo4j_version_is_5_23_or_above = True
     driver.execute_query.return_value = [
         [neo4j_record],
         None,
         None,
     ]
-    search_query, _ = get_search_query(SearchType.HYBRID, return_properties)
+    search_query, _ = get_search_query(
+        SearchType.HYBRID,
+        return_properties,
+        neo4j_version_is_5_23_or_above=retriever.neo4j_version_is_5_23_or_above,
+    )
 
     records = retriever.search(query_text=query_text, top_k=top_k)
 
@@ -355,13 +369,16 @@ def test_hybrid_cypher_retrieval_query_with_params(
         retrieval_query,
         embedder,
     )
+    retriever.neo4j_version_is_5_23_or_above = True
     driver.execute_query.return_value = [
         [neo4j_record],
         None,
         None,
     ]
     search_query, _ = get_search_query(
-        SearchType.HYBRID, retrieval_query=retrieval_query
+        SearchType.HYBRID,
+        retrieval_query=retrieval_query,
+        neo4j_version_is_5_23_or_above=retriever.neo4j_version_is_5_23_or_above,
     )
 
     records = retriever.search(
@@ -419,6 +436,7 @@ def test_hybrid_cypher_retriever_with_result_format_function(
         embedder,
         result_formatter=result_formatter,
     )
+    retriever.neo4j_version_is_5_23_or_above = True
     retriever.driver.execute_query.return_value = [  # type: ignore
         [neo4j_record],
         None,
